Update basket.php
Updated: 
The page now allows the correct data to be inserted into the database creating multiple order_details and linking them to the same order. 
Quantity of users basket now gets updated to the database - this needs to be approved by backend team. 

Due to these quantity selector changes, the user can change the selector on the page without saving to the backend which displays a different total than what it would be as its calculated by input not database. This is a work in progress.