Update basketStyle.css
Updated: 
Minor changes to the css footer. Previously was using bottom:0; to keep the footer to the bottom of the page when the content doesnt fill the screen. However, on smaller screens this stops the user from being able to scroll to see the body content. 
New css has a min-height:100vh; in body and an extra margin-bottom on the checkout-container.